+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add additional information from `GCPCluster` to `cluster-values` configmap when running on `gcp`.
 
+### Changed
+
+- The DNS IP needs to come from the cluster `Services` CIDR.
+- Default `Services` CIDR changed from `172.31.0.0/16` to `10.96.0.0/12` to match k8s default.
+
 ## [1.10.0] - 2022-08-09
 
 ### Changed